+++ src/EntityGraphQL.AspNet/Extensions/AddGraphQLOptions.cs
@@ -3,19 +3,10 @@ using EntityGraphQL.Schema;
 
 namespace EntityGraphQL.AspNet
 {
-    public class AddGraphQLOptions<TSchemaContext>
+    public class AddGraphQLOptions<TSchemaContext> : SchemaBuilderOptions
     {
         /// <summary>
-        /// If true the schema builder will automatically create a singular field for any collection fields whose type has an Id field. 
-        /// E.g. a people field will add a person(id) field.
-        /// </summary>
-        public bool AutoCreateIdArguments { get; set; } = true;
-        /// <summary>
-        /// If true the schema builder will automatically create any Enum types found in the context object graph.
-        /// </summary>
-        public bool AutoCreateEnumTypes { get; set; } = true;
-        /// <summary>
-        /// If true the schema will be built via reflection on the context type.
+        /// If true the schema will be built via reflection on the context type. You can customise this with the properties inherited from SchemaBuilderOptions
         /// If false the schema will be created with the TSchemaContext as it's context but will be empty of fields/types. 
         /// You can fully populate it in the ConfigureSchema callback
         /// </summary>
@@ -23,7 +14,7 @@ namespace EntityGraphQL.AspNet
         /// <summary>
         /// Overwrite the default field naming convention. (Default is lowerCaseFields)
         /// </summary>
-        public Func<string, string> FieldNamer { get; set; } = SchemaBuilder.DefaultNamer;
+        public Func<string, string> FieldNamer { get; set; } = SchemaBuilderSchemaOptions.DefaultFieldNamer;
         /// <summary>
         /// Called after the schema object is created but before the context is reflected into it. Use for set up of type mappings or 
         /// anything that may be needed for the schema to be built correctly.